+++ lib/project/get/get_command.dart
@@ -31,6 +31,10 @@ class GetCommand extends Command {
 
     final yaml = YamlHelper.loadFileYaml(argMorphemeYaml);
 
+    if (yaml.containsKey('repository')) {
+      await 'morpheme repository --morpheme-yaml "$argMorphemeYaml"'.run;
+    }
+
     await ModularHelper.get(concurrent: yaml.concurrent);
     StatusHelper.success('morpheme get');
   }